System to choose answer +Timer for text + Check Correct answer function
Made a system to choose an answer, made a text that appear if the answer chosen is correct or not, add a timer for the text. Note the timer function still doesnt work well dunno what is bug so need work on that plus program is running quite slow sometime need to check on that.